feature modal
Решены все проблемы с модалкой
