+++ src/js/open-modal.js
@@ -4,12 +4,18 @@ import { getMovie } from './movieById';
 
 const filmCardsRef = document.querySelector('.gallery');
 const backdropRef = document.querySelector('.backdrop');
+const modalRef = document.querySelector(".modal");
+const modalCloseBtn = document.querySelector(".modal-btn");
 
 filmCardsRef.addEventListener('click', onFilmCardsClick);
 
 function onFilmCardsClick(e) {
   e.preventDefault();
 
+  document.addEventListener("keydown", onEscTap);
+  backdropRef.addEventListener("click", onClickCloseBtnOrOutside);
+  modalCloseBtn.addEventListener("click", onClickCloseBtnOrOutside);
+  
   if (e.target.nodeName === 'IMG') {
     const movieId = e.target.dataset.id;
 
@@ -25,3 +31,28 @@ async function openModal(movieId) {
   modalCardRender(movie.data);
   // addListenerByBtns(movie);
 }
+
+function onEscTap(e) {
+    
+    if (e.key === "Escape") {
+
+        closeModal()
+    };
+};
+
+function onClickCloseBtnOrOutside(e) {
+
+  if (e.currentTarget.classList.value === "modal-btn" || e.target.classList.value === "backdrop") {
+      
+        closeModal();
+    };  
+};
+
+function closeModal() {
+
+    backdropRef.classList.toggle("is-hidden");
+    document.removeEventListener("keydown", onEscTap);
+    backdropRef.removeEventListener("click", onClickCloseBtnOrOutside);
+    const wrapper = document.querySelector(".wrapper");
+    wrapper.remove();
+};
